requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
moviepy==1.0.3
easyocr==1.7.1
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
SpeechRecognition==3.10.0
boto3==1.34.11
gunicorn==21.2.0
pytest==7.4.3
